tty callback

+++ scripts/build-ffmpeg.sh
@@ -49,8 +49,20 @@ if ! grep -qF "$STDOUT_SCRIPT_FROM" "$WASM_DIR/ffmpeg.js"; then
     exit 1
 fi
 ESCAPED_STDOUT_SCRIPT_FROM=$(printf '%s\n' "$STDOUT_SCRIPT_FROM" | sed -e 's/[]\/$*.^[]/\\&/g');
-ESCAPED_STDOUT_SCRIPT_TO=$(printf '%s\n' "$STDOUT_SCRIPT_TO" | sed -e 's/[]\/$*.^[]/\\&/g');
+ESCAPED_STDOUT_SCRIPT_TO=$(printf '%s\n' "$STDOUT_SCRIPT_TO" | sed -e 's/[\/&]/\\&/g');
 sed -i -e "s/$ESCAPED_STDOUT_SCRIPT_FROM/$ESCAPED_STDOUT_SCRIPT_TO/g" $WASM_DIR/ffmpeg.js
+
+# replacing tty write script in ffmpeg.js
+TTY_SCRIPT_FROM="try{for(var i=0;i<length;i++){stream.tty.ops.put_char(stream.tty,buffer[offset+i])}}"
+TTY_SCRIPT_TO="try{if(Module.tty && Module.tty(stream,buffer,offset,length,pos))var i=length;else for(var i=0;i<length;i++){stream.tty.ops.put_char(stream.tty,buffer[offset+i])}}"
+if ! grep -qF "$TTY_SCRIPT_FROM" "$WASM_DIR/ffmpeg.js"; then
+    echo "ffmpeg.js does not contain expected TTY_SCRIPT_FROM"
+    exit 1
+fi
+ESCAPED_TTY_SCRIPT_FROM=$(printf '%s\n' "$TTY_SCRIPT_FROM" | sed -e 's/[]\/$*.^[]/\\&/g');
+ESCAPED_TTY_SCRIPT_TO=$(printf '%s\n' "$TTY_SCRIPT_TO" | sed -e 's/[\/&]/\\&/g');
+sed -i -e "s/$ESCAPED_TTY_SCRIPT_FROM/$ESCAPED_TTY_SCRIPT_TO/g" $WASM_DIR/ffmpeg.js
+
 gzip --force -9 -c $WASM_DIR/ffmpeg.js > $WASM_DIR/ffmpeg.js.gz
 rm $WASM_DIR/ffmpeg.js
